测得33135&33136两个问题;

+++ SMG_NothingIsAll/AIFoundation/AIThinkingControl/Util/Model/HEResult.h
@@ -31,6 +31,6 @@
 -(HEResult*) mkSPScore:(CGFloat)spScore;
 -(CGFloat) spScore;
 -(HEResult*) mkPStrong:(NSInteger)value;
--(NSInteger) pStrong;
+-(CGFloat) pStrong;
 
 @end
